(PC-10775) move logic to get displayed exclu modules before
